+++ src/util/pelz_request_handler.c
@@ -28,7 +28,7 @@ RequestResponseStatus pelz_encrypt_request_handler(RequestType request_type, cha
       return SIGNATURE_ERROR;
     }
   }
-  int index;
+  size_t index;
   
   unsigned char* cipher_name_string = null_terminated_string_from_charbuf(cipher_name);
   if(cipher_name_string == NULL)
@@ -167,7 +167,7 @@ RequestResponseStatus pelz_decrypt_request_handler(RequestType request_type, cha
   }
   
   charbuf plain_data_internal;
-  int index;
+  size_t index;
 
   unsigned char* cipher_name_string = null_terminated_string_from_charbuf(cipher_name);
   if(cipher_name_string == NULL)
@@ -237,7 +237,7 @@ RequestResponseStatus file_encrypt_in_enclave(charbuf plain_data, charbuf cipher
   pelz_sgx_log(LOG_DEBUG, "File Encryption");
   key->len = 32;
   ocall_malloc(key->len, &key->chars);
-  if(RAND_priv_bytes(key->chars, key->len) != 1)
+  if(RAND_priv_bytes(key->chars, (int)key->len) != 1)
   {
     pelz_sgx_log(LOG_DEBUG, "Key generation failed");
     return ENCRYPT_ERROR;
